correções de sintaxe

+++ JogadorSimples.java
@@ -2,10 +2,10 @@ public class JogadorSimples extends Jogador{
     public JogadorSimples(){
         super(2,6);
     }
-    public void setAtk(short atk){
+    public void setAtk(int atk){
         this.atk = atk;
     }
-    public void setDef(short def){
+    public void setDef(int def){
         this.atk = def;
     }
 }
